[01. 08 - 김진녕] 1. 상세보기 데이터 - 뷰 바인딩
