+++ app/src/main/java/io/sinzak/android/remote/dataclass/product/Product.kt
@@ -3,6 +3,7 @@ package io.sinzak.android.remote.dataclass.product
 import com.google.gson.annotations.SerializedName
 import io.sinzak.android.system.LogError
 import io.sinzak.android.utils.DateFormats
+import io.sinzak.android.utils.TimeUtil
 import java.lang.Exception
 import java.text.NumberFormat
 import java.util.*
@@ -35,28 +36,7 @@ data class Product(
     }
 
     fun getTimePassed() : String {
-        try {
-            val time = DateFormats.MySqlFormat.parse(date.toString().replace('T', ' '))
-            val timeGap = (System.currentTimeMillis() - time.time)/ 1000L / 3600
-
-
-
-            when{
-                timeGap >= 365 * 24 ->
-                    return "· ${(timeGap / (365 * 24)).toInt()}년전"
-                timeGap >= 24 * 30->
-                    return "· ${(timeGap / 24 / 30).toInt()}개월 전"
-                timeGap >= 24 ->
-                    return "· ${(timeGap / 24).toInt()}일 전"
-                else ->
-                    return "· ${timeGap.toInt()}시간 전"
-            }
-
-        } catch (e: Exception)
-        {
-
-        }
-        return "· 최근"
+        return TimeUtil.getTimePassed(date.toString())
     }
 
     fun getFormattedDate() : String{
