+++ app/src/main/java/io/sinzak/android/ui/main/market/artdetail/ContentViewModel.kt
@@ -10,10 +10,19 @@ class ContentViewModel @Inject constructor(
     val model : MarketProductModel
 ): BaseViewModel(){
 
-
+    val art get() = model.art
 
     val imgAdapter = VpAdapter()
 
+    init{
 
+        invokeStateFlow(model.art){art->
+            art?.let{
+                imgAdapter.imgs = it.imgUrls
+                imgAdapter.notifyDataSetChanged()
+            }
+
+        }
+    }
 
 }
