+++ app/src/main/res/layout/view_artdetail_content.xml
@@ -1,6 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout>
 
+    <data>
+        <variable
+            name="vm"
+            type="io.sinzak.android.ui.main.market.artdetail.ContentViewModel" />
+    </data>
+
     <androidx.appcompat.widget.LinearLayoutCompat
 
         android:layout_marginTop="15dp"
@@ -16,7 +22,7 @@
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="24dp"
 
-            android:text="나의 뭐시기를 표현한 작품. 캔버스 10호. 아크릴.택배 거래 안돼요, 어쩌구 저쩌구. 홍대입구에서 직거래 선호합니다. "
+            android:text="@{vm.art.content}"
             android:lineSpacingMultiplier="1.2"
             style="@style/medium16"/>
 
@@ -29,7 +35,7 @@
             >
             <TextView
 
-                android:text="10"
+                android:text="@{String.valueOf(vm.art.wishCnt)}"
 
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -39,7 +45,7 @@
 
             <TextView
 
-                android:text="10"
+                android:text="@{String.valueOf(vm.art.views)}"
                 android:layout_marginStart="5dp"
 
                 android:layout_width="wrap_content"
@@ -50,7 +56,7 @@
 
             <TextView
 
-                android:text="10"
+                android:text="@{String.valueOf(vm.art.chatCnt)}"
 
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
